+++ chronicle_cls/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.4.1
+## Added
+- Added support for region selection.
+
+## Fixed
+- Removed invalid UDM fields from the default mapping file.
+
 # 1.3.1
 ## Changed
 - Mapped 'security_result.action_details' UDM field to Netskope field 'action'.